transaction-service/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
